# Optimize ERB string concat using opt_ltlt